Conflicts:
	build/cmake/CondorConfigure.cmake
	src/condor_daemon_core.V6/daemon_core.cpp
	src/condor_starter.V6.1/vanilla_proc.cpp

+++ src/condor_procd/procd_ctl.cpp
@@ -439,9 +439,9 @@ get_usage(ProcFamilyClient& pfc, int argc, char* argv[])
 	}
 #endif
 	if (pfu.block_read_bytes >= 0)
-		printf("Bytes read from block devices (KB): %lu\n", pfu.block_read_bytes/1024);
+		printf("Bytes read from block devices (KB): %llu\n", (unsigned long long)(pfu.block_read_bytes/1024));
 	if (pfu.block_write_bytes >= 0)
-		printf("Bytes written to block devices (KB): %lu\n", pfu.block_write_bytes/1024);
+		printf("Bytes written to block devices (KB): %llu\n", (unsigned long long)(pfu.block_write_bytes/1024));
 	return 0;
 }
